test: LottoManager.test.js 에서 calculateLotto.test.js로 분리
